start/resume and buildings
StartNewCharacter.php is the script that allows players to create a new
user entry, or restart their old entry with a new player.

resumecharachter.php retrieves the matching player data using the
player id, and return a json object with the game data for that player.

the newbuildingcleared.php is meant to create an entry with a playerid,
bldg_id, and updated timestamp.  My plan is to create a chron_script
that runs every so often and makes the entries “active” again.  I will
also write a query script for that database, that sets cleared
buildings to inactive in the buildingpopulator.cs