Home base database added
added the homebase_sheet to track players home stats.  Mobile client,
when in range of home, can drop supply into the homebase db.

Much of the early systems for the home base client have a start in this
commit.  I’m going to be working on that client further this week.